+++ src/lib/compareBcrypt.ts
@@ -1,0 +1,5 @@
+import { compare } from 'bcrypt';
+
+const compareBcrpyt = (plain: string, hashed: string): Promise<boolean> => compare(plain, hashed);
+
+export default compareBcrpyt;