+++ src/test/bcrypt.test.ts
@@ -1,15 +1,16 @@
-import { compare, hash } from 'lib/bcrypt';
+import compareBcrpyt from 'lib/compareBcrypt';
+import hashBcrypt from 'lib/hashBcrypt';
 
 describe('bcrypt', () => {
     test('Generates a string', () => {
-        hash('foo').then(res => {
+        hashBcrypt('foo').then(res => {
             expect(typeof res).toBe('string');
         });
     });
     test('Successful compare', async () => {
         const data = 'Two tiny timid toads trying to trot to Tarrytown.';
-        const hashed = await hash(data);
-        await expect(compare(data, hashed)).resolves.toBe(true);
-        await expect(compare(hashed, 'foo')).resolves.toBe(false);
+        const hashed = await hashBcrypt(data);
+        await expect(compareBcrpyt(data, hashed)).resolves.toBe(true);
+        await expect(compareBcrpyt(hashed, 'foo')).resolves.toBe(false);
     });
 });